requirements.txt
smtplib
random
time
os
logging
MIMEMultipart
MIMEText
MIMEApplication
load_dotenv